requirements.txt
playwright>=1.58.0
pytest>=9.0.0
pytest-playwright>=0.7.2
allure-pytest>=2.15.3
pyyaml>=6.0.3
Faker>=25.0.0
pytest-html>=4.2.0
tenacity>=9.0.0
